requirements.txt
asgiref==3.8.1
Django==5.1.3
gunicorn==23.0.0
packaging==24.2
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
